+++ bin/check-es-query-exists.rb
@@ -157,10 +157,11 @@ class ESQueryExists < Sensu::Plugin::Check::CLI
          short: '-u USER',
          long: '--user USER'
 
-  option :header,
-         description: 'Headers to pass to the elasticsearch http client.',
-         short: '-H header',
-         long: '--header header'
+  option :headers,
+         description: 'A comma separated list of headers to pass to elasticsearch http client',
+         short: '-H headers',
+         long: '--headers headers',
+         default: 'Content-Type: application/json'
 
   option :timeout,
          description: 'Elasticsearch query timeout in seconds',